% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/context_feature.R
\name{context_feature}
\alias{context_feature}
\title{First line is title}
\usage{
context_feature(data, col, win, bind = T, ...)
}
\arguments{
\item{...}{explanation of \code{...}}

\item{x}{explanation of \code{x}}
}
\value{
The value returned by this function.
}
\description{
This is a brief description.
}
\details{
Anything else after the description goes to the Details section.

You can write several paragraphs.
}
\examples{
split_filename('foo.bar')
}
\author{
Who are you?
}
\references{
\url{https://github.com/yihui/rmini}
}
\seealso{
\code{\link[tools]{file_ext}}, \code{\link[tools]{file_path_sans_ext}}
}

