% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crf_test.R
\name{crf_test}
\alias{crf_test}
\title{First line is title}
\usage{
crf_test(test_data, model_path = "model.out", nbest = "0",
  cost_factor = "1", label = F)
}
\arguments{
\item{x}{explanation of \code{x}}

\item{...}{explanation of \code{...}}
}
\value{
The value returned by this function.
}
\description{
This is a brief description.
}
\details{
Anything else after the description goes to the Details section.

You can write several paragraphs.
}
\examples{
split_filename('foo.bar')
}
\author{
Who are you?
}
\references{
\url{https://github.com/yihui/rmini}
}
\seealso{
\code{\link[tools]{file_ext}}, \code{\link[tools]{file_path_sans_ext}}
}

