% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genia_tagger.R
\name{genia_tagger}
\alias{genia_tagger}
\title{First line is title}
\usage{
genia_tagger(file = "/Users/lipidong/Desktop/genia.txt", named_entity = T)
}
\arguments{
\item{x}{explanation of \code{x}}

\item{...}{explanation of \code{...}}
}
\value{
The value returned by this function.
}
\description{
This is a brief description.
}
\details{
Anything else after the description goes to the Details section.

You can write several paragraphs.
}
\examples{
split_filename('foo.bar')
}
\author{
Who are you?
}
\references{
\url{https://github.com/yihui/rmini}
}
\seealso{
\code{\link[tools]{file_ext}}, \code{\link[tools]{file_path_sans_ext}}
}

